replace some specific function don't work in linux

+++ include/Object.h
@@ -10,9 +10,9 @@ class Object
 {
 public:
     Object();
-    void* operator new(unsigned int size) throw();
+    void* operator new(size_t size) throw();
     void operator delete(void* p);
-    void* operator new[](unsigned int size) throw();
+    void* operator new[](size_t size) throw();
     void operator delete[](void* p);
     bool operator ==(const Object& obj);  //这里实现的是地址相等与否的比较，即是否是同一个对象
     bool operator !=(const Object& obj);
