+++ src/Object.cpp
@@ -7,7 +7,7 @@ Object::Object()
 
 }
 
-void* Object::operator new(unsigned int size) throw()
+void* Object::operator new(size_t size) throw()
 {
     return malloc(size);
 }
@@ -17,7 +17,7 @@ void Object::operator delete(void* p)
     free(p);
 }
 
-void* Object::operator new[](unsigned int size) throw()
+void* Object::operator new[](size_t size) throw()
 {
     return malloc(size);
 }
